refactor: extract date formatting to the separate method
